added google analytics and other fixes

+++ src/com/liorginsberg/lgtasks/TaskDB.java
@@ -97,7 +97,7 @@ public class TaskDB {
 			
 			
 			for(c.moveToFirst(); !c.isAfterLast(); c.moveToNext()) {
-				tasksList.add(new Task(c.getInt(iRow), c.getString(iTitle), c.getString(iDesc), c.getString(iFrom), c.getString(iTo), c.getString(ilocation), c.getInt(iChecked)));
+				tasksList.add(new Task(c.getLong(iRow), c.getString(iTitle), c.getString(iDesc), c.getString(iFrom), c.getString(iTo), c.getString(ilocation), c.getInt(iChecked)));
 			}
 		} catch (SQLException e) {
 			Toast.makeText(context, "Could Not open database error", Toast.LENGTH_LONG).show();
